+++ src/Akeneo/Connectivity/Connection/back/tests/Integration/Apps/Persistence/DbalConnectedAppRepositoryIntegration.php
@@ -166,7 +166,7 @@ FROM akeneo_connectivity_connected_app
 WHERE id = :id
 SQL;
 
-        $row = $this->connection->fetchAssoc($query, [
+        $row = $this->connection->fetchAssociative($query, [
             'id' => $id,
         ]);