+++ src/Akeneo/Platform/Bundle/CatalogVolumeMonitoringBundle/Persistence/Query/Sql/CountAttributes.php
@@ -20,16 +20,12 @@ class CountAttributes implements CountQuery
     /** @var Connection */
     private $connection;
 
-    /** @var int */
-    private $limit;
-
     /**
      * @param Connection $connection
      */
-    public function __construct(Connection $connection, int $limit)
+    public function __construct(Connection $connection)
     {
         $this->connection = $connection;
-        $this->limit = $limit;
     }
 
     /**
@@ -42,7 +38,7 @@ class CountAttributes implements CountQuery
             FROM pim_catalog_attribute;
 SQL;
         $result = $this->connection->executeQuery($sql)->fetchAssociative();
-        $volume = new CountVolume((int) $result['count'], $this->limit, self::VOLUME_NAME);
+        $volume = new CountVolume((int) $result['count'], self::VOLUME_NAME);
 
         return $volume;
     }