+++ src/Akeneo/Platform/Bundle/CatalogVolumeMonitoringBundle/Persistence/Query/Sql/CountProductValues.php
@@ -20,17 +20,12 @@ class CountProductValues implements CountQuery
     /** @var Connection */
     private $connection;
 
-    /** @var int */
-    private $limit;
-
     /**
      * @param Connection $connection
-     * @param int        $limit
      */
-    public function __construct(Connection $connection, int $limit)
+    public function __construct(Connection $connection)
     {
         $this->connection = $connection;
-        $this->limit = $limit;
     }
 
     /**
@@ -44,7 +39,7 @@ class CountProductValues implements CountQuery
 SQL;
         $result = $this->connection->executeQuery($sql)->fetchAssociative();
 
-        $volume = new CountVolume((int) $result['sum_product_values'], $this->limit, self::VOLUME_NAME);
+        $volume = new CountVolume((int) $result['sum_product_values'], self::VOLUME_NAME);
 
         return $volume;
     }