+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/public/js/controller/catalog-volume.tsx
@@ -5,7 +5,7 @@ import {pimTheme} from 'akeneo-design-system';
 import React from 'react';
 import * as $ from 'jquery';
 import BaseView = require('pimui/js/view/base');
-import {CatalogVolumeMonitoringApp} from '@akeneo-pim-community/catalog-volume-monitoring';
+import {CatalogVolumeMonitoringApp, getCatalogVolume} from '@akeneo-pim-community/catalog-volume-monitoring';
 
 const mediator = require('oro/mediator');
 const featureFlags = require('pim/feature-flags');
@@ -21,7 +21,7 @@ class CatalogVolumeController extends ReactController {
     return (
       <DependenciesProvider>
         <ThemeProvider theme={pimTheme}>
-          <CatalogVolumeMonitoringApp />
+          <CatalogVolumeMonitoringApp getCatalogVolumes={getCatalogVolume} />
         </ThemeProvider>
       </DependenciesProvider>
     );