+++ tests/back/Platform/Acceptance/CatalogVolumeMonitoring/Context/LocaleContext.php
@@ -37,16 +37,6 @@ final class LocaleContext implements Context
     }
 
     /**
-     * @Given the limit of the number of locales is set to :limit
-     *
-     * @param int $limit
-     */
-    public function theLimitOfTheNumberOfLocalesIsSetTo(int $limit): void
-    {
-        $this->inMemoryQuery->setLimit($limit);
-    }
-
-    /**
      * @Then the report returns that the number of locales is :numberOfLocales
      *
      * @param int $numberOfLocales
@@ -56,25 +46,5 @@ final class LocaleContext implements Context
         $volumes = $this->reportContext->getVolumes();
 
         Assert::eq($numberOfLocales, $volumes['count_locales']['value']);
-    }
-
-    /**
-     * @Then the report warns the users that the number of locales is high
-     */
-    public function theReportWarnsTheUsersThatTheNumberOfLocalesIsHigh(): void
-    {
-        $volumes = $this->reportContext->getVolumes();
-
-        Assert::true($volumes['count_locales']['has_warning']);
-    }
-
-    /**
-     * @Then the report does not warn the users that the number of locales is high
-     */
-    public function theReportDoesNotWarnTheUsersThatTheNumberOfLocalesIsHigh(): void
-    {
-        $volumes = $this->reportContext->getVolumes();
-
-        Assert::false($volumes['count_locales']['has_warning']);
     }
 }