+++ tests/back/Platform/Acceptance/CatalogVolumeMonitoring/Context/LocalizableAndScopableAttributeContext.php
@@ -37,16 +37,6 @@ final class LocalizableAndScopableAttributeContext implements Context
     }
 
     /**
-     * @Given the limit of the number of localizable and scopable attributes is set to :limit
-     *
-     * @param int $limit
-     */
-    public function theLimitOfTheNumberOfLocalizableAndScopableAttributesIsSetTo(int $limit): void
-    {
-        $this->inMemoryQuery->setLimit($limit);
-    }
-
-    /**
      * @Then the report returns that the number of localizable and scopable attributes is :number
      *
      * @param int $number
@@ -56,25 +46,5 @@ final class LocalizableAndScopableAttributeContext implements Context
         $volumes = $this->reportContext->getVolumes();
 
         Assert::eq($number, $volumes['count_localizable_and_scopable_attributes']['value']);
-    }
-
-    /**
-     * @Then the report warns the users that the number of localizable and scopable attributes is high
-     */
-    public function theReportWarnsTheUsersThatTheNumberOfLocalizableAndScopableAttributesIsHigh(): void
-    {
-        $volumes = $this->reportContext->getVolumes();
-
-        Assert::true($volumes['count_localizable_and_scopable_attributes']['has_warning']);
-    }
-
-    /**
-     * @Then the report does not warn the users that the number of localizable and scopable attributes is high
-     */
-    public function theReportDoesNotWarnTheUsersThatTheNumberOfLocalizableAndScopableAttributesIsHigh(): void
-    {
-        $volumes = $this->reportContext->getVolumes();
-
-        Assert::false($volumes['count_localizable_and_scopable_attributes']['has_warning']);
     }
 }